修改代码，增加状态维护

+++ src/service_database/database_business/entity/model_util_handle/model_status_information_entity.rs
@@ -10,7 +10,7 @@ pub struct Model {
     pub model_id: String,
     pub name: String,
     pub code: String,
-    pub update_at: i32,
+    pub update_at: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
